// libs/core/src/lib/models/bilingual-text.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

/**
 * The model class to hold English and Arabic values of a field.
 *
 * @export
 * @class BilingualText
 */
export class BilingualText {
  english: string = undefined;
  arabic: string = undefined;
}